fix `do` for `StateT`
Because `StateT` lazily evaluates the functions passed into `flat_map`,
the ultimate `send` that results in `StopIteration` produces a duplicate
application of the penultimate result, as that one is returned from the
`do` iteration into the last `flat_map`.

If this is `StateT.modify`, the function is applied twice.

This change calls `Monad.pure` on the passed in value instead of
returning the previous effect instance.